The "Checkbox" and "PhoneInput" components have been moved to "common/components" since they will be reused in the future in different parts of the entire project and not just within the components for which they were originally created. For the same reason, the "Checkmark" component was moved to the "common/icons" directory.
