+++ common/components/checkbox/checkbox.module.css
@@ -4,6 +4,8 @@
   justify-content: center;
   width: 1.125rem;
   height: 1.125rem;
+  min-width: 1.125rem;
+  min-height: 1.125rem;
   position: relative;
 }
